think through hydration
