+++ protocol-designer/src/file-data/selectors/commands.js
@@ -8,7 +8,7 @@ import uniqBy from 'lodash/uniqBy'
 import type {BaseState, Selector} from '../../types'
 import {getAllWellsForLabware} from '../../constants'
 import * as StepGeneration from '../../step-generation'
-import {selectors as steplistSelectors} from '../../steplist/reducers'
+import {selectors as steplistSelectors} from '../../steplist'
 import {selectors as pipetteSelectors} from '../../pipettes'
 import {selectors as labwareIngredSelectors} from '../../labware-ingred/reducers'
 import type {Labware} from '../../labware-ingred/types'
