+++ protocol-designer/src/navigation/selectors.js
@@ -5,7 +5,7 @@ import type {BaseState, Selector} from '../types'
 import {rootSelector as navigationRootSelector} from './reducers'
 import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
 import wellSelectionSelectors from '../well-selection/selectors'
-// import {selectors as steplistSelectors} from '../steplist/reducers'
+// import {selectors as steplistSelectors} from '../steplist'
 
 import type {Page} from './types'
