+++ protocol-designer/src/steplist/formLevel/warnings.js
@@ -5,34 +5,37 @@ import {getWellTotalVolume} from '@opentrons/shared-data'
 ** Warning Messages **
 ********************/
 
-// TODO: reconcile difference between returning error string and key
-
-export type FormWarning = 'OVER_WELL_VOLUME_MAXIMUM'
-
-const FORM_ERRORS: {[FormError]: string} = {
+export type FormWarning = 'OVER_MAX_WELL_VOLUME'
+const FORM_WARNINGS: {[FormWarning]: string} = {
   OVER_MAX_WELL_VOLUME: 'Dispense volume will overflow a destination well'
 }
+export type warningChecker = (mixed) => ?string
 
 // TODO: test these
 /*******************
 ** Warning Checkers **
 ********************/
 
-type maxWellVolumeParams = {labware?: ?string, wells?: ?Array<string>, volume?: ?number}
-export const maxWellVolume = ({labware, wells, volume}: maxWellVolumeParams): ?string => {
+type maxWellVolumeFields = {'dispense--labware'?: ?string, 'dispense--wells'?: ?Array<string>, volume?: ?number}
+export const maxWellVolume = (fields: maxWellVolumeFields): ?string => {
+  // const labware = fields['dispense--labware']
+  // const wells = fields['dispense--wells']
+  const {volume} = fields
+  if (!wells) return null
   wells.forEach(well => {
-    if (volume > getWellTotalVolume(labware, well)) return FORM_ERRORS.OVER_MAX_WELL_VOLUME
+    const maximum = getWellTotalVolume(labware, well)
+    console.log('warnings', labware, wells, volume, maximum)
+    if (maximum && (volume > maximum)) return FORM_WARNINGS.OVER_MAX_WELL_VOLUME
   })
-  return null
 }
 
 /*******************
 **     Helpers    **
 ********************/
 
-// export const composeErrors = (...errorCheckers: Array<errorChecker>) => (value: mixed): Array<string> => (
-//   errorCheckers.reduce((accumulatedErrors, errorChecker) => {
-//     const possibleError = errorChecker(value)
-//     return possibleError ? [...accumulatedErrors, possibleError] : accumulatedErrors
-//   }, [])
-// )
+export const composeWarnings = (...warningCheckers: Array<warningChecker>) => (formData: mixed): Array<string> => (
+  warningCheckers.reduce((acc, checker) => {
+    const possibleWarning = checker(formData)
+    return possibleWarning ? [...acc, possibleWarning] : acc
+  }, [])
+)
