+++ protocol-designer/src/steplist/index.js
@@ -1,6 +1,7 @@
 // @flow
 // TODO Ian 2018-01-23 use this index.js for all `steplist` imports across PD
-import rootReducer, {selectors} from './reducers'
+import rootReducer from './reducers'
+import selectors from './selectors'
 import * as actions from './actions'
 import * as utils from './utils'
 import {getFieldErrors, processField} from './fieldLevel'
