+++ protocol-designer/src/types.js
@@ -4,7 +4,7 @@ import type {RootState as FileData} from './file-data'
 import type {RootState as LabwareIngred} from './labware-ingred/reducers'
 import type {RootState as Navigation} from './navigation'
 import type {RootState as Pipettes} from './pipettes'
-import type {RootState as StepList} from './steplist/reducers'
+import type {RootState as StepList} from './steplist'
 import type {RootState as WellSelection} from './well-selection/reducers'
 export type BaseState = {
   dismiss: Dismiss,
